uploaderScript-v2/requirements.txt
Package                  Version
------------------------ ---------
cachetools               5.3.1
certifi                  2023.7.22
charset-normalizer       3.2.0
google-api-core          1.34.0
google-api-python-client 1.12.8
google-auth              2.23.2
google-auth-httplib2     0.1.1
googleapis-common-protos 1.60.0
httplib2                 0.22.0
idna                     3.4
pip                      20.2.3
protobuf                 3.20.3
pyasn1                   0.5.0
pyasn1-modules           0.3.0
pyparsing                3.1.1
requests                 2.31.0
rsa                      4.9
setuptools               49.2.1
six                      1.16.0
uritemplate              3.0.1
urllib3                  2.0.5
